refactor: remove unecessary type conversions

+++ server/core/player_handler.go
@@ -65,7 +65,7 @@ func (h *PlayerHandler) onPlayerUpdate(player *model.Player, c *websocket.WSConn
 	return func(buf []byte) {
 		msg := &protobuf.Player{}
 		proto.Unmarshal(buf, msg)
-		lat, lon := float64(float32(msg.GetLat())), float64(float32(msg.GetLon()))
+		lat, lon := msg.GetLat(), msg.GetLon()
 		if lat == 0 || lon == 0 {
 			return
 		}
